move .yt.vis subaccessor up to .yt
