+++ yt_xarray/tests/test_yt_wrappers.py
@@ -20,7 +20,7 @@ def xr_ds():
 
 @pytest.mark.parametrize("viz_method", ["SlicePlot", "ProjectionPlot"])
 def test_2d_volume_plots(tmp_path, xr_ds, viz_method):
-    func = getattr(xr_ds.yt.vis, viz_method)
+    func = getattr(xr_ds.yt, viz_method)
     slc = func("x", "temperature")
 
     output_dir = tmp_path / "output"
@@ -33,7 +33,7 @@ def test_2d_volume_plots(tmp_path, xr_ds, viz_method):
 
 
 def test_phase_plot(tmp_path, xr_ds):
-    slc = xr_ds.yt.vis.PhasePlot("pressure", "temperature", "temperature")
+    slc = xr_ds.yt.PhasePlot("pressure", "temperature", "temperature")
 
     output_dir = tmp_path / "output"
     output_dir.mkdir()
@@ -45,7 +45,7 @@ def test_phase_plot(tmp_path, xr_ds):
 
 
 def test_profile_plot(tmp_path, xr_ds):
-    slc = xr_ds.yt.vis.ProfilePlot("pressure", "temperature", weight_field="precip")
+    slc = xr_ds.yt.ProfilePlot("pressure", "temperature", weight_field="precip")
 
     output_dir = tmp_path / "output"
     output_dir.mkdir()
